+++ Changes.md
@@ -12,6 +12,11 @@ Fixes
 - RenderMan : Fixed interactive denoising when rendering via the `it` display driver.
 - EditScopeUI : Fixed crash if `childNodesAreReadOnly` metadata was edited while no EditScope was selected.
 
+API
+---
+
+- Arnold ShaderNetworkAlgo : Deprecated substitutions API. Use IECoreScene's render adaptor API instead. For backwards compatibility until it is removed, the substitutions API forwards to the render adaptor API.
+
 1.6.8.0 (relative to 1.6.7.0)
 =======